orderModel 金額改為double 型態

+++ API/KKday.API.WMS/KKday.API.WMS/Models/DataModel/Booking/OrderModel.cs
@@ -14,8 +14,8 @@ namespace KKday.API.WMS.Models.DataModel.Booking
         public DateTime order_date { get; set; }//訂購日期
         public string order_type { get; set; }//訂單類型(b2b)
         public string order_status { get; set; }//訂單狀態
-        public int order_amt { get; set; }//總金額
-        public int order_b2c_amt { get; set; }//直客總金額
+        public double order_amt { get; set; }//總金額
+        public double order_b2c_amt { get; set; }//直客總金額
         public string connect_name { get; set; }//訂購人姓名
         public string connect_tel { get; set; }//訂購人聯絡電話
         public string connect_mail { get; set; }//訂購人聯絡信箱
@@ -71,8 +71,8 @@ namespace KKday.API.WMS.Models.DataModel.Booking
         public int cus_seqno { get; set; }//旅客細項編號
         public string prod_no { get; set; }//商品編號
         public string prod_name { get; set; }//商品名稱
-        public int prod_amt { get; set; }//商品金額
-        public int prod_b2c_amt { get; set; }//直客商品金額
+        public double prod_amt { get; set; }//商品金額
+        public double prod_b2c_amt { get; set; }//直客商品金額
         public string prod_currency { get; set; }//商品金額幣別
         public int discount_xid { get; set; }//折扣xid
         public string prod_cond1 { get; set; }//人(price1,price2)
